+++ what4/src/What4/Expr/GroundEval.hs
@@ -44,7 +44,6 @@ import           Control.Monad
 import           Control.Monad.Trans.Class
 import           Control.Monad.Trans.Maybe
 import qualified Data.BitVector.Sized as BV
-import           Data.List.NonEmpty (NonEmpty(..))
 import           Data.Foldable
 import qualified Data.Map.Strict as Map
 import           Data.Maybe ( fromMaybe )
@@ -320,15 +319,7 @@ evalGroundApp f a0 = do
       if xv then f y else f z
 
     NotPred x -> not <$> f x
-    ConjPred xs ->
-      let pol (x,Positive) = f x
-          pol (x,Negative) = not <$> f x
-      in
-      case BM.viewBoolMap xs of
-        BM.BoolMapUnit -> return True
-        BM.BoolMapDualUnit -> return False
-        BM.BoolMapTerms (t:|ts) ->
-          foldl' (&&) <$> pol t <*> mapM pol ts
+    ConjPred cm -> BM.evalConj f cm
 
     RealIsInteger x -> (\xv -> denominator xv == 1) <$> f x
     BVTestBit i x ->